pyproject.toml
[tool.poetry]
name = "skillos"
version = "0.1.0"
description = "SkillOS MVP scaffold"
authors = ["Alexey"]
readme = "README.md"
packages = [{ include = "skillos" }]

[tool.poetry.scripts]
skillos = "skillos.cli.main:main"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5"
click = "^8.1"
asyncpg = "^0.29"
redis = "^5.0"
pyyaml = "^6.0"
python-dotenv = "^1.0"
httpx = "^0.25"
structlog = "^24.1"
fastapi = "^0.110"
uvicorn = "^0.27"
psycopg = {version = "^3.1", extras = ["binary"]}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"
ruff = "^0.1"
black = "^23.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
